Additions to documentation and configs
Removed unused modules from requirements.txt
  cupshelpers
  louis
  lxml
  pycups
  pycurl
  pygobject
  team

Added nginx.conf as an example of how to configure uwsgi with nginx.
This can be dropped in as a replacement for pxe nginx config file as
it includes both locations.

Added isic_uwsgi.service file to be placed in /etc/systemd/service/
to start the uwsgi app.

Added SUBDIR to settings.py so that the uwsgi app responds to
requests when it is served under the /icis subdirectory.

Signed-off-by: John Andersen <john.s.andersen@intel.com>