Changed localhost:5000 to 192.168.1.1/icis/
static/ister/* all contained localhost:5000 as the location of
ister/icis configs. This would never be the case as to the machine
installing localhost is just a pxe booted installer image which is
definitely not hosting those configs. Based on the network_boot
documentation and the bulk_provisioning documentation this should
instead be http://192.168.1.1/icis/.

Signed-off-by: John Andersen <john.s.andersen@intel.com>