Creating install scripts for PXE and ICIS
These install scripts set up a PXE server with the ICIS service.  They automate the configuration of Clear Linux for baremetal installations of Clear Linux.  They assume a dedicated machine and a clean install is the starting point.  They also assume that a NAT network topology is used for PXE booting.  They create parameters which users can modify prior to installation.  The built-in parameterized systemd service and socket for uwsgi applications is being used to deploy ICIS on Clear Linux.