Put nginx configurations in dedicated directory
The default configuration of nginx was updated to use a configuration
directory.  This allows multiple applications to be hosted on the same
nginx instance/host as long as different port numbers are used.
Updating the install script to use this change so that users can get
going with IPXE without problems.